@Bouncner 2021-01-26
Not part of this PR, but do you know where the tasks in TPC-C come from? #2287 should have gotten rid of most spawned tasks in TPC-C. Just curious if we have have missed one operator and the same changes could be an easy win here.