@mrks 2021-01-26
Nope. I haven't had the chance to look into TPC-C for a while. :/

> Besides that: humongous results! Is there any wisdom you can share how you came up with that solution or rather how you found the initial issue?

Mmh, it started with you wondering about the nanosleep syscall. Tried to understand why there were threads waiting if there was obviously enough work to do. I'll try do some parallel construction on how it went from there. Remind me to talk about it tomorrow, would you? Mostly reading code and some intution, tbh.

>  Is there a proper way to analyse the scheduler? Might be helpful for upcoming generations.

Not really. I'm not even sure what that would look like. Talking about the scheduler is probably a good idea though. Maybe we should track all the things that you want to talk about as part of the handover somewhere?