@Bouncner 2021-01-28
I still have to get my around this problem and this issue. What bothers me is the TPC-C regression. Not that I care too much about it, but I'd like to ensure that we do not introduce problems that are currently only visible in TPC-C, but might hurt us in the long run.