@mrks 2021-01-28
Here's my current explanation: The "bug" in the current master causes stuck clients. If a single worker keeps multiple clients occupied, the degree of parallelism is reduced. For TPC-C, this is a good thing, as it leads to fewer MVCC conflicts.