@Bouncner 2021-01-29
But shouldn't the rate of failed transactions change then?

```
+--------------++----------+---------+--------++----------+----------+--------+----------------------+
 | Item         || Latency (ms/iter)  | Change || Throughput (iter/s) | Change |              p-value |
 |              ||      old |     new |        ||      old |      new |        |                      |
 +--------------++----------+---------+--------++----------+----------+--------+----------------------+
-| Delivery     ||    131.4 |   141.2 |   +7%  ||     7.26 |     6.74 |   -7%  | (run time too short) |
 |    unsucc.:  ||      4.0 |     3.7 |   -7%  ||   133.80 |   133.68 |   -0%  |                      |
-| New-Order    ||     69.8 |    77.0 |  +10%  ||   121.50 |   110.67 |   -9%  |               0.0000 |
 |    unsucc.:  ||      4.0 |     3.9 |   -2%  ||  1465.16 |  1469.03 |   +0%  |                      |
 | Order-Status ||      6.3 |     6.5 |   +4%  ||   141.04 |   140.43 |   -0%  | (run time too short) |
-| Payment      ||     11.2 |    12.4 |  +11%  ||    10.70 |     8.21 |  -23%  | (run time too short) |
 |    unsucc.:  ||      3.7 |     3.7 |   -2%  ||  1505.67 |  1501.46 |   -0%  |                      |
 | Stock-Level  ||     17.5 |    17.4 |   -1%  ||   141.02 |   140.38 |   -0%  |               0.4717 |
 +--------------++----------+---------+--------++----------+----------+--------+----------------------+
-| Sum          ||    236.2 |   254.5 |   +8%  ||          |          |        |                      |
-| Geomean      ||          |         |        ||          |          |   -8%  |                      |
 +--------------++----------+---------+--------++----------+----------+--------+----------------------+
 ```