@mrks 2021-01-29
> #2287 should have gotten rid of most spawned tasks in TPC-C.

Most, but not all:

```
Start OperatorTask with id: 4294967295 for op: TableScan Impl: Unset OL_O_ID = 2101
Start {Task with id: 4294967295}
Finish {Task with id: 4294967295}
Start {Task with id: 4294967295}
Finish {Task with id: 4294967295}
Finish OperatorTask with id: 4294967295 for op: TableScan Impl: ColumnVsValue OL_O_ID = 2101
```

(No new info here, just sharing a finding)