@mrks 2021-01-29
I gave this some more thought, but I would have to start investigating the TPC-C in much more detail for that. Right now, I do not have the time for that. You raise an absolutely valid point, but I am not sure how to properly address it right now. Once the single-threaded performance of the TPC-C transactions stops being multiple orders of magnitude from any reasonable performance, we would have to revisit the scheduler anyway.