@mrks 2021-01-29
Read: Let's look into the optimizer issues, the indexing, etc. first.