@Bouncner 2021-02-11
Any ideas of this clang fail has something to do with TPC-H Q6?