@mrks 2021-02-15
> Any ideas of this clang fail has something to do with TPC-H Q6?

Unlikely. I changed the system meta table in a previous version of this PR. Somewhere, something got stuck. If this persists, I will wait for this PR to be approved and then re-open it as a new PR.